Standardise chunk store tests and check errors too
Create a base testing class with a single test that checks `put` and
`get` and `ChunkNotFound` on all chunk stores in the same way. It is
a single test as this caters for a reasonably onerous setup() and
teardown() - think FakeS3. For tests checking `StoreUnavailable`,
rather create separate 'dud' testing classes that create straw chunk
stores which trigger this error.

Finally, fix some errors in RadosChunkStore exposed by these tests
(yay!).