Construct RadosChunkStore from available ioctx
The original constructor now becomes the `from_config` class method,
while the default initialiser takes a pre-configured rados.Ioctx object
as parameter. This is more flexible, allowing full customisation
of the RADOS setup while retaining the convenience of `from_config` for
everyday use.

Fix S3ChunkStore to also check for botocore in the client-driven
initialiser, since the EndpointConnectionError in the error map
will only exist if botocore is installed.