Improve RADOS error report if chunk is too big
With RADOS it is necessary to specify the number of bytes when reading
an object. If the object is smaller than expected, only the actual
number of bytes will be read, which is fine. If the object is larger
than expected, however, it is more cumbersome to determine the actual
size without invoking a second call to the object store.

The current solution is to always try to read one byte more than
expected. This is sufficient to detect an error without reading a
potentially very large object and without always calling the object
store twice. Once the specific error is confirmed, though, it is useful
to do a stat() on the object and get its actual size without reading
the whole thing, so that the BadChunk error report makes more sense.