Don't look at CBF for the centre frequency
In the early days, CBF always reported 1284 MHz as the centre frequency
in SPEAD item `center_freq`, stored as attribute `cbf_center_freq` in
both TelescopeModel and later TelescopeState. Then they changed their
minds and made it 428 MHz always. Now it is back to 1284 MHz (probably
due to CAM).

It therefore cannot be trusted and should be ignored - that is what
the hard-coded band-based workaround in H5DataV3 is for. Unfortunately
the latest stream attr machinery from #105 re-opened that wound.
Specifically for center_freq, look in streams and SDP but never in the
old CBF locations. Also only look in telstate for CBF attributes if
it can also look in TelescopeModel for it, as these locations should
have the same info anyway.

This closes katdal GitHub issue #108.