Refactor attribute search to improve clarity
Remove `telstate` from search helper method names as they can search
TelescopeModel as well. Also be specific about which stream applies:

    _get_telstate_cbf_attr    -> _get_cbf_attr
    _get_telstate_stream_attr -> _get_l0_attr

The main improvement is to add yet another intermediate helper method,
:meth:`_raise_if_attr_found`, which constructs an attribute name from
various parts, searches for the attribute either in TelescopeModel or
TelescopeState and raises an exception containing the value if found.
This makes it easy to chain a series of search locations with decreasing
priority. It also logs the results of the search, for debug purposes.

Rearrange the method parameters to minimise verbosity when called.
Convert `stream_name` HDF5 attr to string (it's Unicode on Python 2).

These improvements should make it easier to debug attribute issues
like #108 in the future.