Switch to class-level fixtures in ChunkStore tests
This sets up and tears down an appropriate ChunkStore only once
per class, in order to save time. This also gets rid of the dud
versions of each test class, as these "store unavailable" tests can
now be included in the main class without any overhead. The main
motivation is the pending tests for dask functionality.

The only hold-out is DictChunkStore, which is very fast to instantiate
anyway. I am using a vars() trick to figure out the arrays in the store
and switching from an instance to a class changes the output of this
function in an unpleasant way.