Add dask support to ChunkStore
Add `get_dask_array` method that combines all the chunks associated
with an array into a single dask Array and loads the associated data
only on compute(). The corresponding `put_dask_array` method stores
the chunks of a dask array in the store and returns a results array
that also suppresses any exceptions (for inspection by the caller).
The supporting `generate_chunks` function provides a standard way
to split an array into roughly equal-sized chunks. Dask is only a
test requirement for now - wait until we have a fully integrated
v4 loader to make it a main requirement.