Small improvements to generate_chunks
Remove the visibility-specific default value for `dims_to_split` and
replace it with the most generic option: split on all dimensions if
possible. Add a `power_of_two` option that forces chunk sizes to be
powers of two on those dimensions that support it (i.e. ones with a
power-of-two length).