Improved power-of-two chunks and random tokens
The `generate_chunks` function now allows chunks to have a power-of-two
size even if the parent dimension has a size that is not a power of two.
In this case the target chunk size will be rounded down to the nearest
power of two and the relevant dimension will be divided into N-1 blocks
of this size with the last block taking up the slack and being
potentially smaller than the rest. This is the behaviour of
`toolz.partition_all` (compare with the alternative behaviour of
`np.array_split`, where the blocks are as similar as possible).

The output array name in `put_dask_array` is now mildly randomised
to avoid name clashes and cache reuse in the dask graph. There should
be very few repeats of the same array_name anyway, so the full 32-byte
UUID string is truncated to 8 bytes for readability in the graph when
debugging. If this ever becomes an issue, just use the full string
instead.