Support chunk offsets in dask array put and get
This adds an optional offset to the chunk indices when addressing them
in the store. The main purpose is to put a smaller array into the store
as part of a bigger one (useful for dealing with split heaps in the
visibility writer). The functionality is also added to the getter to
simplify testing (as we are not really dealing with missing data yet).