Add ChunkStore.has_chunk
This checks whether the specified chunk exists in the store with the
appropriate size (and dtype if known). The idea is that it should be
quicker than the full `get_chunk`, which remains the fallback option.

For NPY files we read the header only (could even be reduced to a
filename check if speed is an issue). For RADOS we do ioctx.stat.
For S3 we do a HEAD check, but list_objects will be faster for the
higher-level multi-chunk checks. This also required a botocore
exception remapping.