Remove hacky botocore error remap
When doing `head_object` as opposed to `get_object`, the HTTP request
only returns the headers and not the body, which leads to different
exception handling in botocore. Instead of trying to ram it through
_standard_errors, do a custom try...except with the generic ClientError
but check the response code to detect a missing key and/or bucket.
This is OK since we generally swallow this exception anyway, so the
added benefit of _standard_errors (chunk identification in the error
message) falls away.