Remove timestamp checks
The v4 timestamps are effectively synthesised from dump indices
as part of the chunk info at the end of vis writing. It is therefore
impossible that the number of dumps in the timestamps can differ from
the shape of the vis data as they are generated together. The delta
between timestamps will also be an integer number of dumps, so the
nefarious "Irregular timestamps detected" warning is no more.