Move construction from config file contents
Revert the changes in commit 174ccaced at the class level and rather
move it up to the caller's context. The problem is that a file name
and file contents are both strings. If a file name is intended but the
file does not exist, RadosChunkStore would attempt to interpret it as
file contents, with a cryptic error message as result.

The seek(0) is unnecessary because we are passing the file name and
not the actual pointer. It is necessary to flush the file to the OS,
though, otherwise Rados() can't read it.

This addresses comments in PR #118.