Autodetect program block ID and stream name
The default TelstateDataSource constructor now takes a telstate that
already has views associated with the capture block and stream. This
requires the reintroduction of `chunk_name` to figure out the chunk
store name of the dataset (as `capture_stream` is now implicit again).
This is the intended way to open pipeline data.

When constructing telstate from an URL with no prior views, detect
the capture block ID and stream name. Crash if there are none and warn
if there are more than one. Then prepare the telstate with the relevant
views before passing on to the usual constructor.