Support v3.9 files
The latest version of HDF5v3 (presumably live on 2018-02-21) contains
some elements of v4. The obs sensors live in a capture_block_id
namespace in telstate, and obs_params is now a dict-valued attribute
instead of a sensor of 'key repr(value)' strings. This is indicated
by the presence of 'capture_block_id' in the top-level file attrs.