Synthesise correlator timestamps from telstate
Since katsdpingest PR #229 it is now possible to synthesise the
correlator timestamps based on the relevant telstate keys
('<stream>_sync_time', '<stream>_int_time' and
'<cbid>_<stream>_first_timestamp') and the number of dumps.

This has a number of advantages. The data set can be interpreted
without resorting to the chunk store, simplifying metadata extraction.
The timestamps chunk was a small silly thing to begin with. Synthesis
also enables more accurate timestamps as more bits are available for
each timestamp.