Shorten telstate keys used for sensors
This strips off the most relevant prefix from the start of each
telstate key to get the corresponding sensor name (i.e. the first
prefix in the list that fits the key). A key like `<cbid>_obs_label`
would hereby turn into a sensor called `obs_label`, given the standard
telstate views.

This is not really the best way of doing this. In future I would
maintain a lightweight link to the underlying telstate within
SensorCache and let it do the name mangling itself, at the cost of
not having all the sensor names available in a list up front.