Make writing an NPY file atomic
This writes to a temporary filename first (ending in '.writing.npy')
before renaming it to the final '.npy' file after the file has been
closed. This ensures that other processes waiting for the file (like
the vis trawler) knows when it is safe to pick it up.

Also fix a grammar bug.