Don't make any assumptions on NPY version of chunk
This checks for all known NPY file versions, not just the default
"None" auto-version, just in case we decide to always write v2 files,
for example. The other reason for this is to avoid calling an
underscored internal function in NumPy (who knows why it is underscored
in the first place?).

This whole check might be short-lived once the `dtype` parameter
disappears (and higher level checks could bypass this too).