Improve autodetection of chunk store
For RDB files this first attempts to locate an NpyFileChunkStore in
an adjacent directory named after the capture stream, as is typically
produced by vis_writer and other tools. Failing this, it falls back
to the S3 gateway suggested by telstate. Another fail here is treated
as an error (as this is typically due to authentication or network
connection errors that the user wants to know about). The redis server
option only attempts the S3 option, since there is no file access.

If the user really wants no chunk store it can be specified as None,
which is useful for metadata extraction.

Also update the S3ChunkStore docs after the changes of the previous PR.