Remove old h5toms.py and objtoms.py in favour of mvftoms.py
h5toms.py is turned into a few lines of wrapper to use mvftoms, with a
warning note - to avoid breaking people's existing workflows. objtoms.py
is removed on the assumption that there are probably few enough places
its used that it is easy to update them. The content of mvftoms is all
pushed into a main function so that it can be called from h5toms.py.