Improve UVW coordinate calculation in mvftoms
The coordinates are now in the reference frame of the array centre,
rather than of each of the individual antennas, which should be more
accurate (for some arbitrary file the changes were up to 3.5e-7
relative, so it won't make much difference).

More usefully, the computation is now much faster, because the expensive
parts only scale with the number of antennas rather than the number of
baselines, and some spherical geometry is possibly eliminated because
the antenna positions are already specified in ENU relative to the array
centre.