Combine baseline permutation and transposition into one
This removes the need for a separate intermediate array.

Averaging used to occur between these steps; it's been moved before.
That means averaging will now be done on autocorrelations even if they
aren't needed in the final result. I'm ignoring the performance impact
of that because I plan to optimise the averaging later.