Pipelined MS writing
The writes to the measurement set are written by a separate
multiprocessing process. The theory is that this should allow the
writing to overlap with the loading.

On my laptop it isn't getting any speedup. That might be because there
is still a serialisation to flush at the end of each scan, or it might
just be that this isn't the bottleneck.