Fix deadlock in child process interaction
The Python docs for subprocess repeatedly warns about the dangers of
stderr=PIPE and its propensity to deadlock if not combined with the
communicate() method. And sure enough, we have deadlock (funnily enough
on macOS but not on Linux, probably due to buffering differences).

We cannot use communicate() as the stderr output only becomes available
after the process exits. We need to access it early to find the
allocated port number in order to talk to the fakes3 server.

Redesign the interaction to include a stderr consumer thread (which
also scrapes the port number) and a queue to pass the port number back
to the main thread.

Also reduce the fakes3 timeouts to 1 second (in case something goes
wrong) and simplify the store_unavailable test by using the timeout
parameter in S3ChunkStore.from_url.