Suppress fakes3 stdout output
This is a belated supplement to PR #150. It suppresses the stdout of
the fakes3 child process, since nose doesn't capture it and it then
obscures the usual test output. Capturing stdout via a pipe instead is
cumbersome and we don't need it (except maybe for debugging, in which
case we can temporarily undo this patch).

For more discussion on this limitation of nose, see
https://stackoverflow.com/q/23679431.