Replace missing chunks with zeros in dask array
Add `get_chunk_or_zeros` method that swallows `ChunkNotFound` and
returns an array of zeros of the appropriate shape and dtype. This is
the appropriate behaviour for vis, weights and flags, so make it a
generic feature.

Next, make this the default underlying getter in `get_dask_array`.
Missing chunks can still be identified via `has_dask_array`, but
neither methods will now raise `ChunkNotFound`.

Test this in all stores except DictChunkStore, which comes prepopulated
with chunks and therefore cannot have missing chunks (if accessed via
the usual tests that preserve chunk structure).