Flag missing chunks as "data_lost"
When accessing the flags, do a `ChunkStore.has_dask_array` on each
of the `correlator_data`, `weights` and `weights_channel` arrays and
flag any missing chunks as "data_lost" (i.e. bit 3 of the flag byte).

The hard-coded value for "data_lost" will soon be resolved when the
flag info is moved to a higher level.