Slight tweak to broadcasting of weights_channel
Instead of making another dask array with `chunks_lost[..., np.newaxis]`
let the broadcasting happen inside the _multi_or function (which now
forces everything to be 3-D). The relevant block index 'ijk' becomes
'ij' in this case, which alerts da.atop that `weights_channel` is
different.

There seems to be no difference in performance but there are fewer
lines of code (and one fewer dask layer).