Don't check chunk size in has_chunk
Make `has_chunk` as lightweight as possible. It now only checks the
presence or absence of a chunk based on its name (i.e. it checks
whether an NPY file or S3 object exists). The chunk dtype and shape
will still be checked by `get_chunk` and result in a `BadChunk`
exception if wrong. There is no need to anticipate this in `has_chunk`.

This fixes NumPy 1.14 support, which modified the size of the NPY file
header by changing its padding (see numpy/numpy#8598). This made it
harder to predict the expected S3 object size (which was still being
produced by an older NumPy), so get rid of the check.