Optimise has_dask_array
Since the flags now call `has_dask_array` on the rest of the arrays
in the dataset, it is time to optimise it. This feature can also be
used to check which chunks are present upon opening a dataset.

Some chunk store implementations allow an efficient enumeration of
all chunks associated with an array. We are only interested in the
presence or absence of a chunk and nothing more. For example, the
`NpyFileChunkStore` can `os.listdir` the array subdirectory to get
a list of all the relevant NPY files, while the `S3ChunkStore` can
send a `list_objects_v2` request with the appropriate bucket and prefix
to get a list of associated keys (via pagination).

If this efficient `list_chunk_ids` method is available, use it within
`has_dask_array`, otherwise fall back to a series of `has_chunk` calls.
Introduce the `chunk_id_str` method to simplify generation of chunk
identifier strings.