Refactor optimised has_dask_array into has_array
Extracting an ndarray of bools from `has_dask_array` can be quite slow
(i.e. 7 seconds for an array with 27000 chunks). To make things worse,
the optimised version first produces an ndarray, which is then chunked
into tiny bits. The fastest route is therefore to return this ndarray
as an end product for those who want it, via the new `has_array` method.
The dask version is moved to `has_dask_array` instead. This allows the
other "has_chunk" private version of `has_dask_array` to be folded back
into its parent. This provides a cleaner and faster API.