Fix MS basename
The base name of the MS was lost in the mix and crashes the --uvfits
(-u) option. Restore it and also use it for the caltable section.
Include some general cleanup like deleting commented-out lines of code
and simplifying string formatting and path name manipulation.

This addresses JIRA ticket COMM-315.