Don't warn if flags selection is empty string
All flags can be discarded by specifying an empty sequence or an empty
string to select()'s flags parameter. The latter resulted in a warning
that '' is not a legitimate flag type, though, because the comma
splitter created [''] instead of []. Suppress this warning and treat
'' and [] the same.

This addresses JIRA ticket COMM-316.