Enforce timeout when constructing S3 chunk store
On initialisation of S3ChunkStore a connection is made to the S3
endpoint URL to check that it is up. While useful, this can lead to
interminable delays while doing DNS lookups for archive-gw-1.kat.ac.za
(at least from my admittedly pathological home setup via sshuttle).

The worst part is that DNS lookup via socket.gethostbyname() ignores
the usual timeout mechanisms. Therefore introduce a hard cutoff by
running the blocking initialisation code in a separate thread. The
constructed store (or an exception) is returned via a queue. This
smacks of concurrent.futures but avoids adding yet another dependency
on Python 2 just for this (revisit when Python 3 only...).

Test that the initialisation happens within the expected time period.