Only produce metadata-only dataset if asked
Instead of automatically falling back to a metadata-only dataset in
the absence of a chunk store, do a hard crash so that the user knows
that visibility data is not coming. By passing `chunk_store=None` to
katdal.open() the user promises not to access vis and friends, which
then also bypasses all chunk store functionality. This is the
sanctioned explicit way to open a metadata-only dataset.