Use relevant dtype and shape for empty selections
When categorical sensor values are looked up at an empty selection of
dumps (e.g. as `sensor[[]]`) the returned empty array defaults to a
dtype of float64 and shape of (0,). This causes problems when
concatenated with other (presumably non-empty) arrays of the correct
dtype in a ConcatenatedDataSet. The default shape is also an issue
when concatenating array-like values like cal products.

Figure out the appropriate dtype and shape in the special case of
an empty output array, by doing a stripped-down version of
:attr:`CategoricalData.dtype` that also peers inside opaque array-like
values and picks the underlying dtype.

This addresses JIRA ticket SR-1193.