Make Subarray and SpectralWindow hashable
These objects are stored as sensors even though they are technically
attributes of a dataset, mostly to enable concatenation of dissimilar
data sets. When the sensors are concatenated these objects are compared
by identity (via a dict) and not by value. This resulted in failure
to concatenate similar datasets, thereby defeating the whole purpose
of concatenation.

Fix this by making Subarray and SpectralWindow hashable so that
similar objects will become equal by identity. Use a similar approach
to katpoint.Target and friends by defining a description property
(string or tuple will do). The description is not meant to reconstruct
or serialise the object (yet?) so rather make it private. Arrange the
fields of the descriptions to enable a sensible ordering of objects
(especially for SpectralWindow, where the window with the lowest
frequencies should come first).

As a side benefit this allows the culling of the modified `array_equal`
function which only served comparisons of Subarrays and SpectralWindows.