Clarify use of integer division in freqs formula
This is not really a fix since all MeerKAT datasets to date have had an
even number of channels and katdal is still Python 2 only, so the new
code behaves identically to the old. This was really done with an eye
towards Python 3, where these distinctions matter more.

It makes more sense to use integer division in the case of an odd
number of channels as this will ensure that the centre frequency is
always in the middle of a specific channel (the central channel even
in this case). It is also more sensible in the case of a single channel.

Do some half-hearted line shortening too.