Incorporate L1 flags into L0 product
If telstate has the 'sdp_archived_streams' key and it lists additional
streams (only in the full RDB), take each stream's 'chunk_info' dict
and let it override the corresponding entries in the default L0 one.
This specifically allows the additional 'sdp_l1_flags' stream to
override the 'flags' array in 'sdp_l0'.

In the process, refactor the chunk_info entries to contain the relevant
prefix aka base_name aka chunk_name, which is the prefix to be added
to the basic array name to find the array in the chunk store. This
makes it trivial to mix and match arrays between arbitrary S3 buckets
and NPY subdirectories. It's a pity this was not the original design
as the 'chunk_name' telstate key could then be chucked...

This addresses JIRA ticket SR-1157.