PR fixes
Only look for additional *flag* streams based on the current stream to
avoid mixing up unrelated streams (e.g. there might be a continuum
stream in `sdp_archived_streams` in the future). Move the flag upgrade
procedure to its own function to clean up the main function.

Set the prefixes of `chunk_info` items to the `chunk_name` telstate key
in a separate function. This allows `chunk_name` to be deprecated in
favour of explicit prefixes in `chunk_info` (to be added to vis_writer
and katsdpflagwriter). Remove all other references to `chunk_name`.