Speed up flagging of lost data
Instead of synthesizing a sequence of masks (which are mostly zero), it
looks at the list of missing chunks, and computes the pieces of each
flag chunk that need the mask applied. A single map_blocks is then used
to apply the mask only where needed. This gives a 10x speedup.

This is not complete, because it requires has_array to be present, and
has_dask_array is now dead code. I also haven't checked test coverage.