Preserve the number of dumps when merging L1 flags
The number of dumps in the sdp_l1_flags stream may differ from that of
the source sdp_l0 stream, since the chunks are produced in separate
unsynchronised writers (flag_writer vs vis_writer) which may experience
different load conditions and dropouts. Consider the sdp_l0 stream
to be the gold standard (it has the visibilities, after all) and extend
or truncate the number of dumps in sdp_l1_flags chunk info to match it.

This addresses JIRA ticket MKAIV-1220.