Eliminate SensorCache.iteritems/itervalues
Testing shows that current versions of Python implement these in terms
of the base dictionary, without calling the overloaded `__getitem__` or
`get`.

This does affect the semantics slightly, because the implementation that
was removed took a copy of the array rather than a view (that gets
broken by modification during iteration).