Support AWS-style authentication to S3ChunkStore
The token-based authentication authenticates with our JWT-based proxy.
This adds credential authentication using an access key and secret key.
It relies on botocore to generate the signatures, which is added as an
optional dependency ("s3credentials").

Also adds a create_array method to ChunkStore, which should be used
before calling put_chunk or put_dask_array to ensure that the bucket
(S3ChunkStore) or directory (NpyFileChunkStore) is created, and that the
bucket policy allows public access if requested.