Attempt at making minio startup work without races
It relies on minio using SO_REUSEADDR so that it can bind to a port at
the same time as our temporary socket is bound to it, which seems to
Work On My Machine (tm). It should be safe since
(a) we don't try to talk to minio until after we've closed the temporary
socket; and
(b) we never call listen() on the temporary socket, so there is no risk
of connections accidentally being directed to it.