Modernise tab completion
Since IPython 5.0 (PR ipython/ipython#9289 specifically) the
customisation of dict key lookups has been considerably simplified by
adding a specific method (`_ipython_key_completions_`) to your object.
For the `SensorCache` it is not even necessary to add this method since
it is derived from the standard dict. Strip out all the old tab
completion machinery as a result.

Unfortunately the rest of the IPython machinery is still a bit clunky
and the default completions now include all known tokens with that
prefix, not only the ones returned by `_ipython_key_completions_`.
This can be fixed by monkeypatching IPython in one of its startup files
(ask the author).