Add mark_complete and is_complete to ChunkStore
This was being handled by katsdpdatawriter, but in a way that would only
work for NpyFileChunkStore. This will allow katsdpdatawriter's internals
to be chunk store agnostic.

It's not implemented for RadosChunkStore, because I don't have a
convenient way to test that. We should consider just deleting
RadosChunkStore since it's never used nor planned to be used.