Switch back to checking minio readiness with HTTP
Minio version 2018-08-25T01:56:38Z fixes minio/minio#6324, so with that
in place it's possible to test for readiness by making HTTP queries
without crashing minio. That allows --quiet to be used again, which
makes it a bit faster since it stops minio checking if there is a newer
version on the download server.

The HTTP check now also pings /minio/health/live, which is an endpoint
designed to tell us if the server is ready to serve, rather than relying
on getting some error back from / when not using credentials.

A check is added to ensure that the minio version includes the bugfix.