Fix advanced indexing semantics of DaskLazyIndexer
This consolidates both the first-stage and second-stage indexing of
dask arrays into a single helper function: `_dask_getitem`.

This function explicitly does outer indexing, aka oindex, thereby
extending the standard dask __getitem__. It does both simplification
of the index expression and culling of unused parts of the resulting
dask graph to improve performance.

A secondary reason to defer the simplify_index step is to let
:attr:`DaskLazyIndexer.keep` reflect the original boolean indices.
This will turn out useful in applycal.

Add a range of unit tests that focus on the oindex semantics of
_dask_getitem, validating it against NumPy where possible and using the
example from NEP 21 as basis. In future this could also be compared
against a proper `ndarray.oindex` member.

Another future enhancement might be to turn the first-stage indexing
into yet another transform, thereby allowing transforms and indexing
to be mingled at will.

This addresses JIRA ticket SR-1323.