New add_transform method on DaskLazyIndexer
The current state of affairs is that all transforms need to be known
before a DaskLazyIndexer is instantiated, as they enter via the
initialiser only. It is not sufficient to append a new one to the
`transforms` list afterwards, as any access to the `dataset` member
will trigger indexing and transformation, whereafter the `transforms`
list is ignored. There is therefore only a small and not very useful
window of opportunity to add a transform.

Improve this by having an explicit `add_transform` member that can
also apply the new transform directly to the current dataset if it
missed the window. This is useful because some transforms
(read: applycal) benefit from the information stored in the
instantiated object (such as chunk info and time/freq/corrprod
selections).