Add __str__ and __repr__ methods for debugging
This follows the lead of the original :class:`LazyIndexer` class.
In the case of __str__, though, only print out the names of each
transform since their effect on shape and dtype are not recorded like
in :class:`LazyTransform`.

This allows transforms to remain simple callables without wrapping them
in a LazyTransform-like class just to get their name. Instead, look for
the function name or something equivalent when __str__ is constructed.
The user can still customise this by setting a __name__ attribute on
the callable.