Add gain correction sensors based on cal product K
Add the first set of gain correction sensors, starting with calibration
pipeline product 'K' (delays). This extracts the relevant cal solutions
from telstate and repackages it as a virtual sensor in the sensor
cache, in a form amenable to applycal. Each sensor is associated with
a single input and expresses the delay as an array of phase terms
(expanded over all the frequency channels of the visibility data).

This serves as skeleton for future correction sensors (B, G, etc).

This addresses JIRA ticket SR-1219.