Add transform for cal corrections to a vis indexer
This is the first attempt to apply calibration corrections within
an indexer, starting with visibilities. This adds a transform which
in turn installs a function (via dask map_blocks) which calibrates
a chunk of visibilities at a time.

For each chunk this function figures out the dumps and channels
involved, gets the relevant gain correction sensor values (per input)
for all requested calibration products, multiplies them together and
expands the correction array over correlation products. It then uses
a user-specified application method to correct the actual data.
For visibilities the corrections are simply multiplied, but for weights
and flags this will differ.

This addresses JIRA ticket SR-1252.