Handle stream inheritance
If a vis or flags stream has an `inherit` member in telstate, the
telstate view will be constructed to have that stream in the telstate
view chain. This supports SR-1348 / ska-sa/katsdpdata#126.

view_capture_stream now just does the work of constructing a telstate
from capture_block_id and stream_name, and the "smart" version is now
view_l0_capture_stream (name chosen since it asserts that the type is
sdp.l0).