Make applycal test more realistic
Instead of ones, use random data for vis. This requires selecting
the appropriate parts of the vis array to multiply with the expected
gains, which opens another can of worms (stage1 + stage2 + oindex).

Select a subset of the correlation products as well, which requires
the expected gains (and applycal transform) to take in the selected
corrprods.