Allow corrprod axis to be chunked in applycal
This avoids checking the chunking of vis/weights/flags to warn
the user, which has the unwanted side effect of premature dask graph
construction. Keep the indexer as lazy as possible...

Other than that, this will have little effect since the fundamental
chunks in katdal typically never split the corrprod axis.