Depend on dask >= 0.18.2
This version of dask introduced the `block_info` keyword for functions
called by :func:`dask.array.map_blocks`, which allows the function to
figure out the location of the incoming chunk within the full array
that contains it. This removes the need for `add_applycal_transform`
to access the indexer dataset chunking scheme upfront, which triggers
construction of a potentially large dask graph before it is needed.

Dask 0.18.2 also allows us to remove a workaround in the lazy indexer
that checks the length of fancy boolean indices before conversion to
integer indices.

Let requirements.txt install the latest dask 0.19.2, which also cleans
up an unaesthetic NumPy issubdtype FutureWarning.