Initial work on supporting weight scaling by autocorr power
It is triggered by a `need_weight_power_scale` telstate key, which won't
exist in any current archive data.

It adds a dependence on numba to achieve high performance. With a read
benchmark on my laptop, I see about 3.5 GB/s without this and 2 GB/s
with it.

It is completely untested.