Fix S3ChunkStore for Python 3.6
While testing on my home machine, I ran into unit test failures because
code inside Python 3.6's httplib does
```
if not chunk:
```
and if chunk is an ndarray then raises ValueError.

Wrap it in a memoryview so that truthiness is more sensible. I haven't
measured the performance impact, but the initial development of the
_Multipart optimisation used a memoryview and performed acceptably
well.