All SpectralWindow to be constructed with a bandwidth
If given, the bandwidth is used to compute channel_width, instead of
vice versa, which may be both more convenient, and more exact (in the
sense of avoiding rounding errors if num_chans isn't a power of 2 times
a divisor of the bandwidth.

This now guarantees that rechannelise preserves bandwidth exactly, but
on the other hand subrange may now change channel_width
infinitestimally if rounding occurs in the calculation of the new
bandwidth (which only occurs if bandwidth * new_channels runs out of
bits).