Handle multi-part cal products
This will be needed for the coming bandpass (B) product. Split the
product sensor extraction out to a new `get_cal_product` function.
This needs access to telstate aka `attrs` to find out whether the
product has multiple parts or not (used so that bmerry won't groan).

An advantage of having attrs around is that other cal attributes (e.g.
ants, pols and freqs) can then be extracted internally in applycal
instead of having them all passed as parameters to the main functions.