Use katsdptelstate.decode_value to decode telstate values
Telstate now provides a decode_value function to decode the stored
values. The values in future may cease to be pickles, so this
abstraction will then be necessary. A nice benefit even before switching
the encoding is that katdal no longer needs to duplicate the magic
incantations necessary to safely decode pickles written by Python 2, and
completely removes the pickle module from katdal.