Fix invalid BP shape and dtype
Ensure that bandpass corrections match the number of channels in the
visibility data (and not in the cal solutions), even if invalid.
Preserve the complex64 dtype when inverting the gains by using
np.reciprocal instead of dividing into float32(1.0).

Use different spectral configurations for the data and cal solutions
in the unit tests to catch out corner cases.