Apply calibration corrections to weights and flags
The weights are divided by the magnitude squared of the corrections.
The flags are set to True wherever the gains were invalid. This is
done after the flags are combined, on the boolean level (as opposed
to the "raw" uint8 level). Ideally a new "cal_failure" flag type
should be introduced instead, but that is left for a future PR.

This addresses JIRA ticket SR-1252.