Synchronise access to cal products in sensor cache
With this lock, all sorts of weird things happen when applying cal on
real data. An example error message on line 131 of applycal.py:

AxisError: axis 1 is out of bounds for array of dimension 1

It turns out that the values returned from the categorical sensors
are wrapped in ComparableArrayWrappers when they shouldn't be. Adding
the lock around get_cal_product makes the problem go away. Mmm.