Remove default reference antenna from v4 data sets
The katdal "reference antenna" (no relation to the calibration ref ant)
is used for three things during construction of a DataSet:

  - It provides the activity sensor that partitions data set into scans
  - It provides the target sensor that identifies all the targets
  - It serves as the default antenna for the DataSet catalogue

The default reference antenna (obtained by setting ref_ant parameter
to '') is currently the first antenna in the list. This is very
problematic if the first antenna misbehaved (see e.g. tickets SR-1383
and COMM-342). In the absence of an explicit reference antenna (the
default case), now do the following:

  - Use the obs_activity sensor to partition the data set (this is what
    the array intends to do, not what individual antennas actually do)
  - Use the cbf_target sensor for targets (this is the delay tracking
    centre, a better array target than individual pointing centres)
  - Use the array reference part of the antenna object as a new default
    catalogue antenna (this is nominally the array centre as antenna)

Picking an explicit ref_ant still recovers the old katdal behaviour.

This addresses JIRA ticket SR-1188.