Expose S3 bucket name on data source
The :class:`ChunkStoreVisFlagsWeights` gains a `prefix` attribute
which corresponds to the S3 bucket name for the visibility data
(considered the main payload and therefore representative of the
data set). This aids the archive metadata extractor by removing the
need to scavenge around in `chunk_info`.

This addresses JIRA ticket SR-1401.